+++ lib/src/utils/dom_selector.dart
@@ -1,43 +1,40 @@
 import '../model/base.dart';
 
 /// Select top element
-XPathNode? top(XPathNode? e) {
+XPathNode<T>? top<T>(XPathNode<T>? e) {
   if (e == null) return null;
-  while (e!.parentNode != null) {
-    e = e.parentNode!;
+  while (e!.parent != null) {
+    e = e.parent!;
   }
   return e;
 }
 
 /// Selects all ancestors (parent, grandparent, etc.) of the current node
-List<XPathNode> ancestor(XPathNode? e) {
-  final result = <XPathNode>[];
+List<XPathNode<T>> ancestor<T>(XPathNode<T>? e) {
+  final result = <XPathNode<T>>[];
   if (e == null) return result;
   var currentDom = e;
 
-  while (currentDom.parentNode != null) {
-    if (currentDom.parentNode! is XPathElement) {
-      result.add(currentDom.parentNode!);
-    }
-    currentDom = currentDom.parentNode!;
+  while (currentDom.parent != null) {
+    currentDom = currentDom.parent!;
   }
   return result;
 }
 
 /// Selects all ancestors (parent, grandparent, etc.) of the current node and the current node itself
-List<XPathNode> ancestorOrSelf(XPathNode? e) {
-  if (e == null) return <XPathNode>[];
-  return <XPathNode>[e, ...ancestor(e)];
+List<XPathNode<T>> ancestorOrSelf<T>(XPathNode<T>? e) {
+  if (e == null) return <XPathNode<T>>[];
+  return <XPathNode<T>>[e, ...ancestor(e)];
 }
 
 /// Selects all children of the current node
-List<XPathNode> child(XPathNode? e) => e?.childrenNode ?? [];
+List<XPathNode<T>> child<T>(XPathNode<T>? e) => e?.children ?? [];
 
 /// Selects all descendants (children, grandchildren, etc.) of the current node
-List<XPathNode> descendant(XPathNode? e) {
-  final result = <XPathNode>[];
+List<XPathNode<T>> descendant<T>(XPathNode<T>? e) {
+  final result = <XPathNode<T>>[];
   if (e == null) return result;
-  for (final child in e.childrenNode) {
+  for (final child in e.children) {
     result.add(child);
     result.addAll(descendant(child));
   }
@@ -45,18 +42,18 @@ List<XPathNode> descendant(XPathNode? e) {
 }
 
 /// Selects all descendants (children, grandchildren, etc.) of the current node and the current node itself
-List<XPathNode> descentOrSelf(XPathNode? e) {
-  if (e == null) return <XPathNode>[];
-  return <XPathNode>[e, ...descendant(e)];
+List<XPathNode<T>> descentOrSelf<T>(XPathNode<T>? e) {
+  if (e == null) return <XPathNode<T>>[];
+  return <XPathNode<T>>[e, ...descendant(e)];
 }
 
 /// Selects everything in the document after the closing tag of the current node
-List<XPathNode> following(XPathNode root, XPathNode? e) {
-  final result = <XPathNode>[];
+List<XPathNode<T>> following<T>(XPathNode<T> root, XPathNode<T>? e) {
+  final result = <XPathNode<T>>[];
   var elementFound = false;
 
-  void dfs(XPathNode parent) {
-    for (final child in parent.childrenNode) {
+  void dfs(XPathNode<T> parent) {
+    for (final child in parent.children) {
       if (child == e) {
         elementFound = true;
       }
@@ -72,25 +69,25 @@ List<XPathNode> following(XPathNode root, XPathNode? e) {
 }
 
 /// Selects all siblings after the current node
-List<XPathElement> followingSibling(XPathElement? e) {
-  final result = <XPathElement>[];
+List<XPathNode<T>> followingSibling<T>(XPathNode<T>? e) {
+  final result = <XPathNode<T>>[];
   if (e == null) return result;
   var currentDom = e;
-  while (currentDom.nextElementSibling != null) {
-    result.add(currentDom.nextElementSibling!);
-    currentDom = currentDom.nextElementSibling!;
+  while (currentDom.nextSibling != null) {
+    result.add(currentDom.nextSibling!);
+    currentDom = currentDom.nextSibling!;
   }
   return result;
 }
 
 /// Selects all siblings before the current node
-List<XPathElement> precedingSibling(XPathElement? e) {
-  final result = <XPathElement>[];
+List<XPathNode<T>> precedingSibling<T>(XPathNode<T>? e) {
+  final result = <XPathNode<T>>[];
   if (e == null) return result;
   var currentDom = e;
-  while (currentDom.previousElementSibling != null) {
-    result.add(currentDom.previousElementSibling!);
-    currentDom = currentDom.previousElementSibling!;
+  while (currentDom.previousSibling != null) {
+    result.add(currentDom.previousSibling!);
+    currentDom = currentDom.previousSibling!;
   }
   return result;
 }