+++ CHANGELOG.md
@@ -18,3 +18,8 @@
 ## 1.1.0
 - Support `not(funcion()|attr = """)`
 - Support namespace function: `local-name()` `name()`
+
+## 2.0.0
+1. Custom parser
+2. Remove class`XPathElement`, which merge to`XPathNode`
+3. In `XPathResult`, `elements`=>`nodes`, `elements`=>`element`